refactor(asr): extract transcriber module, factory, and CLI resolver
Move ConversationWord/ConversationItem, _process_conversation, and the
three Whisper model classes into docling/pipeline/asr_transcriber.py.
Add AsrTranscriber protocol and AsrModelFactory. AsrPipeline now builds
its model via the factory and re-exports moved symbols for backward
compatibility. Extract the CLI ASR enum->preset mapping into
_resolve_asr_options.

No behavior change; existing ASR tests pass.
